# src/api/docs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"StudentDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateStudent": {
				"properties": {
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "nodejs-template",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/students": {
			"post": {
				"operationId": "CreateStudent",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StudentDto"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "John Doe"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create Student Route",
				"tags": [
					"Users Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateStudent"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetStudents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/StudentDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users Routes"
				],
				"security": [],
				"parameters": []
			}
		},
		"/students/{studentId}": {
			"get": {
				"operationId": "GetStudentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StudentDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users Routes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "studentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StudentDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users Routes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "studentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateStudent"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StudentDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users Routes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "studentId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}